// src/api/generated/models/getCompaniesCompanyIdInvoicesInvoiceIdShippableLines200.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { ShippableLine } from './shippableLine';

export type GetCompaniesCompanyIdInvoicesInvoiceIdShippableLines200 = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  invoiceId: string;
  lines: ShippableLine[];
};